refactor: tidy layouts and fix topbar
