+++ web/components/blocks/RightPanelLayout.tsx
@@ -24,9 +24,3 @@ export default function RightPanelLayout({
     </div>
   );
 }
-
-export function RightPanelSkeleton() {
-  return (
-    <div className="p-4 text-sm text-muted-foreground">Right panel ready</div>
-  );
-}
