--- web/components/common/UploadArea.tsx
@@ -1,4 +1,0 @@
-"use client";
-
-export { UploadArea } from "@/components/ui/UploadArea";
-export type { UploadAreaProps } from "@/components/ui/UploadArea";
